+++ CHANGELOG.md
@@ -16,6 +16,9 @@ The binary has the name tt/tarantool_ + seven-digit hash.
 
 - Installation failure from a commit hash.
 - Crash on `tt install <tool> master`.
+- `--with-binaries` flag for `tt pack` not working while packing 
+  with `--cartridge-compat`. 
+- `tarantool` binary after `pack` now always named `"tarantool"`.
 
 ## [1.2.0] - 2023-08-18